# offchain/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the SafeLand project. All endpoints (except /api/frontend/login) require Bearer token authentication.",
        "title": "SafeLand API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/external/gis-extract": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get plot shape GIS data by UPI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Services"
                ],
                "summary": "Get plot shape by UPI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique Parcel Identifier",
                        "name": "upi",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/external/parcel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get parcel information by providing UPI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Services"
                ],
                "summary": "Get parcel by UPI",
                "parameters": [
                    {
                        "description": "Parcel Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ParcelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/external/phoneuser/{phone}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get NID information by phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Services"
                ],
                "summary": "Get NID by phone number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone Number",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/external/tax-arrears": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get tax arrears information for a parcel by UPI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Services"
                ],
                "summary": "Get tax arrears by UPI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique Parcel Identifier",
                        "name": "upi",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/external/title": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get e-title document by UPI and language (returns PDF file)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "External Services"
                ],
                "summary": "Get title by UPI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique Parcel Identifier",
                        "name": "upi",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language (english, french,kinyarwanda etc)",
                        "name": "language",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/external/upis": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all UPIs for an owner. Authenticates with external service and retrieves UPIs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External Services"
                ],
                "summary": "Get UPIs by owner ID",
                "parameters": [
                    {
                        "description": "UPIs Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UPIsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/frontend/login": {
            "post": {
                "description": "Frontend application login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Frontend Auth"
                ],
                "summary": "Authenticate frontend application",
                "parameters": [
                    {
                        "description": "Frontend Login Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.FrontendLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.FrontendLoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ParcelRequest": {
            "type": "object",
            "properties": {
                "owner_id": {
                    "type": "string",
                    "example": "1200480002578150"
                },
                "upi": {
                    "type": "string",
                    "example": "1/02/04/01/2795"
                }
            }
        },
...
        "handlers.UPIsRequest": {
            "type": "object",
            "properties": {
                "id_type": {
                    "type": "string",
                    "example": "NID"
                },
                "owner_id": {
                    "type": "string",
                    "example": "1200480002578150"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}